Merge test and master

+++ app/src/main/java/com/example/nicholastran/quiz/Models/Questions.java
@@ -6,7 +6,7 @@ public class Questions {
 
     public static ArrayList<Questions> questions = new ArrayList<>();
     public static ArrayList<Boolean> userProgress = new ArrayList<>();
-    public static int currentProgress;
+    public static double currentProgress;
 
     private String question;
     private String answer;
